feat(admin-dashboard): restructure Admin Dashboard components and add new features
- Refactored the Admin Dashboard to improve component organization by updating import paths and file structure.
- Introduced new components for Account Management, User Management, and Reports, enhancing user oversight and data analysis capabilities.
- Added styling for new components, including responsive designs for better user experience across devices.
- Implemented LineChart and PieChart components for visual data representation in the Reports section.
- Enhanced AdminProfile and BulkEmail components with additional functionalities and improved UI elements.
