+++ src/pages/AdminDashboard/components/DashBoard/Dashboard.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import LineChart from './LineChart';
+import LineChart from '../Chart/LineChart';
 import './Dashboard.css';
 
 const Dashboard = () => {
